+++ CHANGELOG.md
@@ -12,13 +12,11 @@ All notable changes to **PromptBlender** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.2.5] - 2026-03-05
 
 ### Added
 
 - Ask to save unsaved changes before running combinations
-
-### Changed
 
 ### Fixed